TRS: consolidate posterior/log_likelihood/prior into methods
